Create a separate utility file for the copy and sort function for reusability
