+++ SamplesDashboard/SamplesDashboard/ClientApp/src/utilities/copy-and-sort.ts
@@ -1,0 +1,41 @@
+export function copyAndSort<T>(items: T[], columnKey: string, isSortedDescending?: boolean): T[] {
+    const key = columnKey as keyof T;
+    const itemsSorted = items.slice(0).sort((a: T, b: T) => (compare(a[key], b[key], isSortedDescending)));
+    return itemsSorted;
+}
+
+function compare(a: any, b: any, isSortedDescending?: boolean) {
+    // Handle the possible scenario of blank inputs 
+    // and keep them at the bottom of the lists
+    if (!a) { return 1; }
+    if (!b) { return -1; }
+
+    let valueA: any;
+    let valueB: any;
+    let comparison = 0;
+
+    if (typeof a === 'string' || a instanceof String) {
+        // Use toUpperCase() to ignore character casing
+        valueA = a.toUpperCase();
+        valueB = b.toUpperCase();
+        // its an item of type number
+    } else if (typeof a === 'number' && typeof b === 'number') {
+        valueA = a;
+        valueB = b;
+    } else {
+        // its an object which has a totalCount property
+        valueA = a.totalCount;
+        valueB = b.totalCount;
+    }
+
+    if (valueA > valueB) {
+        comparison = 1;
+    } else if (valueA < valueB) {
+        comparison = -1;
+    }
+
+    if (isSortedDescending) {
+        comparison = comparison * -1;
+    }
+    return comparison;
+}
