+++ SamplesDashboard/SamplesDashboard/ClientApp/src/views/details/Details.tsx
@@ -9,6 +9,7 @@ import { Link } from 'react-router-dom';
 import authService from '../../components/api-authorization/AuthorizeService';
 import PageTitle from '../../components/layout/PageTitle';
 import { IDetailsItem } from '../../types/samples';
+import { copyAndSort } from '../../utilities/copy-and-sort';
 import { filterListClass } from '../repositories/Repositories.Styles';
 import { buttonClass, classNames, descriptionClass, iconClass, linkClass } from './Details.Styles';
 import { RepositoryStatus } from './Details.types';
@@ -93,7 +94,7 @@ export default class Details extends React.Component<any, any> {
         // call the statistics function
         this.statusStatistics(this.allItems);
         this.setState({
-            items: this.sortItems(this.allItems),
+            items: copyAndSort(this.allItems, 'packageName'),
             repositoryDetails: {
             name: repositoryName,
             description: data.description,
@@ -102,11 +103,7 @@ export default class Details extends React.Component<any, any> {
             isLoading: false
         });
     }
-
-    public sortItems(items: IDetailsItem[]) {
-        const sortedItems = items.sort((a, b) => (a.packageName.toLowerCase() > b.packageName.toLowerCase()) ? 1 : -1);
-        return sortedItems;
-    }
+  
     // compute status statistics
     public statusStatistics(items: IDetailsItem[]) {
         let uptoDateCount = 0;
@@ -398,47 +395,6 @@ function checkStatus(status: number)
     }
 }
 
-function copyAndSort<T>(items: T[], columnKey: string, isSortedDescending?: boolean): T[] {
-    const key = columnKey as keyof T;
-    const itemsSorted = items.slice(0).sort((a: T, b: T) => (compare(a[key], b[key], isSortedDescending)));
-    return itemsSorted;
-}
-
-function compare(a: any, b: any, isSortedDescending?: boolean) {
-    // Handle the possible scenario of blank inputs 
-    // and keep them at the bottom of the lists
-    if (!a) { return 1; }
-    if (!b) { return -1; }
-
-    let valueA: any;
-    let valueB: any;
-    let comparison = 0;
-
-    if (typeof a === 'string' || a instanceof String) {
-        // Use toUpperCase() to ignore character casing
-        valueA = a.toUpperCase();
-        valueB = b.toUpperCase();
-        // its an item of type number
-    } else if (typeof a === 'number' && typeof b === 'number') {
-        valueA = a;
-        valueB = b;
-    } else {
-        // its an object which has a totalCount property
-        valueA = a.totalCount;
-        valueB = b.totalCount;
-    }
-
-    if (valueA > valueB) {
-        comparison = 1;
-    } else if (valueA < valueB) {
-        comparison = -1;
-    }
-
-    if (isSortedDescending) {
-        comparison = comparison * -1;
-    }
-    return comparison;
-}
