+++ SamplesDashboard/SamplesDashboard/ClientApp/src/views/details/Details.types.ts
@@ -1,0 +1,3 @@
+export enum RepositoryStatus{
+    unknown, uptoDate, majorUpdate, patchUpdate, urgentUpdate
+}
